Added and Complete "grade_summary-1.sh"
This is a bash script that will hard-code students and their grades in math, science and literature. Then it will average those grades and section it to a letter grade (A, B, C, D, F) and finally print the output with the student's name, average, and grade along with a grade distribution.